Line endings to unix, spaces vs tabs, ] [ => ][
The original files were edited on multiple platforms and didn't have
any particular pressure on them for consistency, so they wound up not
having any.  Go with unix line endings and spaces:

http://blog.hostilefork.com/death-to-carriage-return/

Also tightens brackets when they are the only thing on a line, as it
helps make the code read a little better.