Changes to be able to run in current Ren-C `<r3-legacy>`
When last updated, Draem could run in a circa 2014 R3-Alpha.  It was
not aggressive in using experimental features, because there were no
binary drops with those features readily available...and because it
was considered a good test of a backwards compatibility layer.  It
was also a good way to do a rough comparison of performance by being
able to run the same code in older and newer interpreters.

That compatibility layer had atrophied from lack of use, and no updates
to the blogs using the software were made for 2 years (in part due to
being busy with other things, but largely due to avoiding being taxed
by going over all the details and making decisions about what to do
with the whole thing).

This commit corresponds to a commit bringing the Ren-C experimental
Rebol interpreter up to date w.r.t. its ability to mimic some R3-Alpha
behaviors.  The new goal of this emulation is not so much to mimic
every quirk of Rebol2/R3-Alpha, but to make it *possible* to code in
a compatible subset when the mode is turned on.

Some features are no longer supported.  One of the biggest impacts on
Draem is a feature dropped from Ren-C which few people knew existed in
R3-Alpha at all: the /ONLY form of IF and UNLESS and EITHER, along with
the deprecation of using non-block branches.  The reasoning behind
dropping that is here:

https://forum.rebol.info/t/backpedaling-on-non-block-branches/476

It was a post-open-sourcing change in R3-Alpha, and since it couldn't
be done in Rebol2, there weren't many instances "in the wild".
One replacement option available in Ren-C (though not in the
`<r3-legacy>` emulation) is `??` and `!!` as infix operators.  But
rather than go directly there, this just backpedals and inserts the
blocks where they are needed...even though it adds irritating levels
of indentation in many places.  Alternative approaches to the problems
can be reviewed more holistically in light of newer possibilities.

Most of the changes relate to that, but there also was the removal of
`--`, which was used only once.  It has been supplanted by the infix
ME operator, which is not implementable in R3-Alpha:

https://trello.com/c/8Bmwvwya

Another change is that when blocks are accessed via PATH! and find
nothing, that triggers an error vs. returning a blank.  This is to make
block accesses and object accesses more consistent in their treatment
of missing items, which is relatively important if block-path access
could be expected to run a FUNCTION!, since it silently fails.  While
PICK can be used as the "blankifying" analogue for indices, it won't
work for DATE!, so workarounds are needed...like:

    any [:date/time false]

This gives a consistent result in Rebol2, R3-Alpha, and Ren-C...since
Ren-C will give back a void on missing :date/time instead of a NONE!,
but then omit that as a vote and fall through the false to give BLANK!.