Photobucket is evil and extortive, move to local hosting
Due to laziness I allowed the image links on the site to rot to the
site Photobucket.  Long ago it had been a good place to put images,
but that was very long ago.

These days, I should probably instruct AI about how to illustrate
the environments.  But I scraped the thumbnails of the images I had
stored back in those early days...that's the best I could get.